@@ -9,57 +9,57 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~16.2.0",
-    "@angular/common": "~16.2.0",
-    "@angular/compiler": "~16.2.0",
-    "@angular/core": "~16.2.0",
-    "@angular/forms": "~16.2.0",
-    "@angular/platform-browser": "~16.2.0",
-    "@angular/platform-browser-dynamic": "~16.2.0",
-    "@angular/router": "~16.2.0",
-    "@nx/angular": "16.8.1",
+    "@angular/animations": "^17.0.9",
+    "@angular/common": "^17.0.9",
+    "@angular/compiler": "^17.0.9",
+    "@angular/core": "^17.0.9",
+    "@angular/forms": "^17.0.9",
+    "@angular/platform-browser": "^17.0.9",
+    "@angular/platform-browser-dynamic": "^17.0.9",
+    "@angular/router": "^17.0.9",
+    "@nx/angular": "17.2.8",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~16.2.0",
-    "@angular-devkit/core": "~16.2.0",
-    "@angular-devkit/schematics": "~16.2.0",
-    "@angular-eslint/eslint-plugin": "~16.0.0",
-    "@angular-eslint/eslint-plugin-template": "~16.0.0",
-    "@angular-eslint/template-parser": "~16.0.0",
-    "@angular/cli": "~16.2.0",
-    "@angular/compiler-cli": "~16.2.0",
-    "@angular/language-service": "~16.2.0",
-    "@nx/cypress": "16.8.1",
-    "@nx/eslint-plugin": "16.8.1",
-    "@nx/jest": "16.8.1",
-    "@nx/js": "16.8.1",
-    "@nx/linter": "16.8.1",
-    "@nx/web": "16.8.1",
-    "@nx/workspace": "16.8.1",
-    "@schematics/angular": "~16.2.0",
+    "@angular-devkit/build-angular": "^17.0.10",
+    "@angular-devkit/core": "^17.0.10",
+    "@angular-devkit/schematics": "^17.0.10",
+    "@angular-eslint/eslint-plugin": "^17.0.1",
+    "@angular-eslint/eslint-plugin-template": "^17.0.1",
+    "@angular-eslint/template-parser": "^17.0.1",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.9",
+    "@angular/language-service": "^17.0.9",
+    "@nx/cypress": "17.2.8",
+    "@nx/eslint-plugin": "17.2.8",
+    "@nx/jest": "17.2.8",
+    "@nx/js": "17.2.8",
+    "@nx/web": "17.2.8",
+    "@nx/workspace": "17.2.8",
+    "@schematics/angular": "17.0.10",
     "@tailwindcss/forms": "^0.5.6",
     "@types/jest": "^29.4.0",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1",
+    "@types/node": "18.16.9",
+    "@typescript-eslint/eslint-plugin": "6.20.0",
+    "@typescript-eslint/parser": "6.20.0",
     "autoprefixer": "^10.4.0",
     "cypress": "^13.0.0",
-    "eslint": "~8.46.0",
-    "eslint-config-prettier": "8.1.0",
+    "eslint": "8.48.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-plugin-cypress": "^2.13.4",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
-    "jest-preset-angular": "~13.1.0",
+    "jest-preset-angular": "13.1.6",
     "jsonc-eslint-parser": "^2.1.0",
-    "nx": "16.8.1",
+    "nx": "17.2.8",
     "postcss": "^8.4.5",
     "prettier": "^2.6.2",
     "tailwindcss": "^3.0.2",
     "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "~5.1.3"
+    "typescript": "5.2.2",
+    "@nx/eslint": "17.2.8"
   }
 }
